+++ packages/aws-amplify/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.5](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@4.3.4...aws-amplify@4.3.5) (2021-11-09)
+
+**Note:** Version bump only for package aws-amplify
+
+
+
+
+
 ## [4.3.4](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@4.3.3...aws-amplify@4.3.4) (2021-10-28)
 
 **Note:** Version bump only for package aws-amplify